+++ app/src/main/java/com/monke/monkeybook/presenter/impl/ISearchBookPresenter.java
@@ -4,7 +4,7 @@ package com.monke.monkeybook.presenter.impl;
 import com.monke.basemvplib.impl.IPresenter;
 import com.monke.monkeybook.bean.SearchBookBean;
 
-public interface ISearchPresenter extends IPresenter{
+public interface ISearchBookPresenter extends IPresenter{
 
     Boolean getHasSearch();
 
@@ -26,5 +26,5 @@ public interface ISearchPresenter extends IPresenter{
 
     Boolean getInput();
 
-    void setInput(Boolean input);
+    void upSearchEngineS();
 }
